--- a/src/poly/enc.rs
+++ b/src/poly/enc.rs
@@ -26,7 +26,7 @@ where
     // Use provided scale or calculate half of q
     let scale = M::P::from_const(params, &<M::P as Poly>::Elem::half_q(&params.modulus()));
 
-    // Compute RLWE encryption: t * a + e + m * scale)
+    // Compute RLWE encryption: t * a + e + m * scale
     t.clone() * a + e + &(m.clone() * &scale)
 }