In order to decrypt a ciphertext defined as `a_bit_0, b_bit_0, a_bit_1, b_bit_1, ..., a_bit_logq-1, b_bit_logq-1` the current implementation executes a circuit that performs, for each pair `a_bit, b_bit` the following inner product operation:

`a_bit * t_bar + b_bit * -1` which, overall, accounts for `2*log_q` multiplication gates per ciphertext.

In this proposed PR, the logic has been modified to directly pass `-t_bar` as circuit input and perform the operation `a_bit * -t_bar + b_bit` resulting in `log_q` multiplication gates per ciphertext.

At the same time, we had to modify the sign in the `rlwe_encrypt` logic

from 

```
    b = t * a + e - (m * scale)
```

to 

```
    b = t * a + e + (m * scale)
```

## Benches

Benchmarks calculated on `uv run memory_profile.py cargo test -r --test test_io_middle_param --no-default-features -- --ignored`

`main`

```
Time to obfuscate: 348.691209709s
Time for evaluation: 314.686266083s
Total time: 663.377475792s
```

```
===== PHYSICAL MEMORY USAGE =====

Step Description                                                       Memory Usage    Absolute Change      % Change       
------------------------------------------------------------------------------------------------------------------------
Sampled public data                                                    17.94 MB        0.0 B                0.00%
Sampled pub key init                                                   26.55 MB        8.61 MB              48.00%
Sampled s_bars                                                         26.56 MB        16.00 KB             0.06%
Sampled t_bar_matrix                                                   26.59 MB        32.00 KB             0.12%
Sampled hardcoded_key_matrix                                           26.59 MB        0.0 B                0.00%
Generated RLWE ciphertext {a, b}                                       26.66 MB        64.00 KB             0.24%
Decomposed RLWE ciphertext into {bits(a), bits(b)}                     36.09 MB        9.44 MB              35.40%
Sampled initial encodings                                              45.22 MB        9.12 MB              25.28%
b star trapdoor init sampled                                           76.14 MB        30.92 MB             68.38%
Computed p_init                                                        76.14 MB        0.0 B                0.00%
Computed u_0, u_1, u_star                                              76.14 MB        0.0 B                0.00%
Sampled b_star trapdoor for idx                                        107.41 MB       31.27 MB             41.06%
Sampled pub key idx                                                    109.12 MB       1.72 MB              1.60%
Sampled b trapdoor for idx and bit                                     151.66 MB       42.53 MB             38.97%
preimage parameters computed                                           154.80 MB       3.14 MB              2.07%
p_hat generated                                                        477.88 MB       323.08 MB            208.71%
z_hat_mat generated                                                    501.61 MB       23.73 MB             4.97%
z_hat generated                                                        556.31 MB       54.70 MB             10.91%
Computed m_preimage_bit                                                528.06 MB       -29622272.0 B        -5.08%
preimage parameters computed                                           528.06 MB       0.0 B                0.00%
p_hat generated                                                        630.45 MB       102.39 MB            19.39%
z_hat_mat generated                                                    649.50 MB       19.05 MB             3.02%
z_hat generated                                                        685.47 MB       35.97 MB             5.54%
Computed n_preimage_bit                                                644.61 MB       -42844160.0 B        -5.96%
preimage parameters computed                                           645.92 MB       1.31 MB              0.20%
p_hat generated                                                        798.11 MB       152.19 MB            23.56%
z_hat_mat generated                                                    814.53 MB       16.42 MB             2.06%
z_hat generated                                                        867.44 MB       52.91 MB             6.50%
Computed k_preimage_bit                                                823.84 MB       -45711360.0 B        -5.03%
Sampled b trapdoor for idx and bit                                     825.67 MB       1.83 MB              0.22%
preimage parameters computed                                           826.33 MB       672.00 KB            0.08%
p_hat generated                                                        912.28 MB       85.95 MB             10.40%
z_hat_mat generated                                                    913.94 MB       1.66 MB              0.18%
z_hat generated                                                        941.45 MB       27.52 MB             3.01%
Computed m_preimage_bit                                                916.64 MB       -26017792.0 B        -2.64%
preimage parameters computed                                           916.64 MB       0.0 B                0.00%
p_hat generated                                                        1009.34 MB      92.70 MB             10.11%
z_hat_mat generated                                                    1012.39 MB      3.05 MB              0.30%
z_hat generated                                                        1.03 GB         42.83 MB             4.23%
Computed n_preimage_bit                                                1.01 GB         -19431424.0 B        -1.76%
preimage parameters computed                                           1.01 GB         0.0 B                0.00%
p_hat generated                                                        1.13 GB         122.20 MB            11.79%
z_hat_mat generated                                                    1.14 GB         11.77 MB             1.02%
z_hat generated                                                        1.22 GB         75.92 MB             6.49%
Computed k_preimage_bit                                                1.21 GB         -8601600.0 B         -0.66%
Computed final_circuit                                                 1.19 GB         -16777216.0 B        -1.29%
Evaluated outputs                                                      3.20 GB         2.01 GB              168.16%
Computed final_preimage_target                                         1.32 GB         -2022965248.0 B      -58.86%
preimage parameters computed                                           1.32 GB         0.0 B                0.00%
p_hat generated                                                        1.32 GB         720.00 KB            0.05%
z_hat_mat generated                                                    1.32 GB         0.0 B                0.00%
z_hat generated                                                        1.32 GB         0.0 B                0.00%
Sampled final_preimage                                                 1.32 GB         0.0 B                0.00%

===== SUMMARY =====
Initial physical memory: 17.94 MB
Final physical memory: 1.32 GB
Peak physical memory: 3.20 GB
Total physical memory increase: 1.30 GB
```

`feat/optimize_ip_circuit_2`

```
Time to obfuscate: 274.506186791s
Time for evaluation: 233.976892709s
Total time: 508.4830795s
```

```
===== PHYSICAL MEMORY USAGE =====

Step Description                                                       Memory Usage    Absolute Change      % Change       
------------------------------------------------------------------------------------------------------------------------
Sampled public data                                                    18.11 MB        0.0 B                0.00%
Sampled pub key init                                                   26.86 MB        8.75 MB              48.32%
Sampled s_bars                                                         26.88 MB        16.00 KB             0.06%
Sampled t_bar_matrix                                                   26.88 MB        0.0 B                0.00%
Sampled hardcoded_key_matrix                                           26.88 MB        0.0 B                0.00%
Generated RLWE ciphertext {a, b}                                       26.91 MB        32.00 KB             0.12%
Decomposed RLWE ciphertext into {bits(a), bits(b)}                     36.00 MB        9.09 MB              33.80%
Sampled initial encodings                                              45.08 MB        9.08 MB              25.22%
b star trapdoor init sampled                                           76.25 MB        31.17 MB             69.15%
Computed p_init                                                        76.25 MB        0.0 B                0.00%
Computed u_0, u_1, u_star                                              76.25 MB        0.0 B                0.00%
Sampled b_star trapdoor for idx                                        107.36 MB       31.11 MB             40.80%
Sampled pub key idx                                                    109.03 MB       1.67 MB              1.56%
Sampled b trapdoor for idx and bit                                     150.92 MB       41.89 MB             38.42%
preimage parameters computed                                           152.83 MB       1.91 MB              1.26%
p_hat generated                                                        491.56 MB       338.73 MB            221.64%
z_hat_mat generated                                                    501.78 MB       10.22 MB             2.08%
z_hat generated                                                        544.45 MB       42.67 MB             8.50%
Computed m_preimage_bit                                                518.42 MB       -27295744.0 B        -4.78%
preimage parameters computed                                           519.58 MB       1.16 MB              0.22%
p_hat generated                                                        612.17 MB       92.59 MB             17.82%
z_hat_mat generated                                                    640.89 MB       28.72 MB             4.69%
z_hat generated                                                        673.52 MB       32.62 MB             5.09%
Computed n_preimage_bit                                                630.61 MB       -44990464.0 B        -6.37%
preimage parameters computed                                           631.41 MB       816.00 KB            0.13%
p_hat generated                                                        778.41 MB       147.00 MB            23.28%
z_hat_mat generated                                                    786.62 MB       8.22 MB              1.06%
z_hat generated                                                        851.86 MB       65.23 MB             8.29%
Computed k_preimage_bit                                                830.89 MB       -21987328.0 B        -2.46%
Sampled b trapdoor for idx and bit                                     816.23 MB       -15368192.0 B        -1.76%
preimage parameters computed                                           816.30 MB       64.00 KB             0.01%
p_hat generated                                                        904.31 MB       88.02 MB             10.78%
z_hat_mat generated                                                    907.08 MB       2.77 MB              0.31%
z_hat generated                                                        941.00 MB       33.92 MB             3.74%
Computed m_preimage_bit                                                915.67 MB       -26558464.0 B        -2.69%
preimage parameters computed                                           915.67 MB       0.0 B                0.00%
p_hat generated                                                        1006.28 MB      90.61 MB             9.90%
z_hat_mat generated                                                    1.00 GB         19.92 MB             1.98%
z_hat generated                                                        1.03 GB         28.97 MB             2.82%
Computed n_preimage_bit                                                1.01 GB         -20873216.0 B        -1.89%
preimage parameters computed                                           1.01 GB         0.0 B                0.00%
p_hat generated                                                        1.13 GB         122.56 MB            11.84%
z_hat_mat generated                                                    1.15 GB         24.62 MB             2.13%
z_hat generated                                                        1.21 GB         52.78 MB             4.46%
Computed k_preimage_bit                                                1.20 GB         -10403840.0 B        -0.80%
Computed final_circuit                                                 1.20 GB         0.0 B                0.00%
Evaluated outputs                                                      2.86 GB         1.67 GB              139.19%
Computed final_preimage_target                                         1.29 GB         -1692090368.0 B      -55.06%
preimage parameters computed                                           1.29 GB         0.0 B                0.00%
p_hat generated                                                        1.29 GB         224.00 KB            0.02%
z_hat_mat generated                                                    1.29 GB         0.0 B                0.00%
z_hat generated                                                        1.29 GB         0.0 B                0.00%
Sampled final_preimage                                                 1.29 GB         0.0 B                0.00%

===== SUMMARY =====
Initial physical memory: 18.11 MB
Final physical memory: 1.29 GB
Peak physical memory: 2.86 GB
Total physical memory increase: 1.27 GB
```